Added link to Gantti (A simple PHP Gantt Class)
Added link to Gantti (A simple PHP Gantt Class) in the Miscellaneous section. Though this project's repository has not been commited to since Oct 4, 2012, I still believe this tool is worth including in this list because when I searched packagist (https://packagist.org/search/?q=gantt) using the keyword "gantt", there were only two results (one of which is a link to a gantti repository that cannot be found). Hopefully https://github.com/bastianallgeier/gantti does not later become a dead link. Would it be acceptable if I put the link to the project's homepage (http://bastianallgeier.com/gantti/) instead?