Added Dwoo
Dwoo is another template engine that has Smarty syntax support and has plugin support that can be written with classes (instead like functions in Smarty). Maybe it's not so popular like Smarty, Twig or Blade but I still on Internet is possible to find many lists where this template engine is recommended by someone. So, I think it would be a good idea to add it to the list.