Adds the CakePHP fixture factories
Get rid of static fixtures and insert data on the fly.  The CakePHP fixture factories is a must-have tool for TDD development.

For CakePHP apps, and since v2.5 for any PHP application.

The CakePHP ORM makes it very simple to integrate the present plugin into any PHP application. This already eased the life of quite some CakePHP developers. It is time to share this with the whole PHP community!